Refactor ItemsList component because it was bad.com
