+++ client/src/components/items/index.js
@@ -1,0 +1,4 @@
+import DeleteItem from './DeleteItem';
+import UpdateItem from './UpdateItem';
+
+export { DeleteItem, UpdateItem };
